requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-dotenv==1.0.0
openai~=1.86.0
mangum==0.17.0
pydantic-settings==2.1.0
tenacity==8.2.3
pydantic~=2.11.7
chromadb~=1.0.15
nltk~=3.9.1
requests~=2.32.4
tqdm
beautifulsoup4~=4.13.4
google-generativeai
protobuf~=5.29.5